fix: 修复相关鉴权单机版本运行问题
